feat(language): Add language support
This commit is the start of adding language support to this nvim config.
To enable languages, one can use the new `createNeovimWithLanguages`
function to enable languages and certain language features.
Currently the only supported language is rust.

The default package will remain, as a bare-bones neovim instance,
without support for any languages.